+++ src/anova.c
@@ -1,9 +1,10 @@
 #include "anova.h"
 #include "utils.h"
 #include <stdio.h>
+#include <stdlib.h>
 #include <math.h>
 
-Anova avova_test( groupPtr* group_array, int num_of_groups )
+Anova avova_test( groupPtr* group_array, size_t num_of_groups )
 {
     Anova anova = {0};
 
@@ -45,7 +46,7 @@ Anova avova_test( groupPtr* group_array, int num_of_groups )
     return anova;
 }
 
-void display_anova(Anova anova , groupPtr* group_array, int num_of_groups)
+void display_anova(Anova anova , groupPtr* group_array, size_t num_of_groups)
 {
 
     printf("%s\n\n", "Anova: Single Factor");
